integrated process.py (stub) into CLI
