+++ cli/debug.py
@@ -1,0 +1,4 @@
+from odi import cli
+
+
+cli()
