+++ cli/odi/helpers/config.py
@@ -1,0 +1,16 @@
+import os
+import sys
+
+
+def get(key=None):
+    """Return config dict.
+
+    If key is passed it returns value of key.
+    """
+    base_path = os.path.join(os.getcwd())
+    sys.path.append(base_path)
+    import config_default as config
+    if key:
+        config = getattr(config, key)
+    sys.path.remove(base_path)
+    return config
